commit 446c407deb93d79642a777eef22cf998e34a3248
    10569 MDB module for disk labelling would be useful
commit 9b8f194103574e93a8789067d2afccef3c586885
    10476 file(1) could be smatch clean
commit b6a0e2cd4ce8d91fe9dc750c44e662d7c7e5fb8e
    10366 ld(1) should support GNU-style linker sets
    10581 ld(1) should know kernel modules are a thing
commit ab65fd2a6cf8b63577d26411f2ea4b628591d56d
    10593 illumos build should not use kernel modules as link-editor input
commit c524b4fe42d7c586615aacae917f985e3379a108
    10346 ld(1) should not reduce symbol visibility of COMDAT symbols when producing relocatable objects
commit 32e09e17e4529edf39ffb44fb13cdb6a0fb45733
    10804 Enable unused warnings outside of uts
commit 8e718be9d8b2bb15811a2dc0b61cbb5cc5fa3070
    10770 idm: NULL pointer errors
commit 4da99751f967c5f8f04e17d8881a68c2940f3b06
    10790 i86pc: NULL pointer errors
commit 36f99a58464a4c3e25e881cab429881edb09951b
    10778 mac: NULL pointer errors

	Conflicts:
	usr/src/uts/common/io/mac/mac_datapath_setup.c
	usr/src/uts/common/io/mac/mac_client.c
	usr/src/uts/common/io/mac/mac.c

+++ usr/src/uts/common/io/idm/idm_so.c
@@ -388,7 +388,7 @@ idm_v6_addr_okay(struct in6_addr *addr6)
 int
 idm_get_ipaddr(idm_addr_list_t **ipaddr_p)
 {
-	ksocket_t 		so4, so6;
+	ksocket_t		so4, so6;
 	struct lifnum		lifn;
 	struct lifconf		lifc;
 	struct lifreq		*lp;
@@ -622,7 +622,7 @@ idm_iov_sosend(ksocket_t so, iovec_t *iop, int iovlen, size_t total_len)
 {
 	struct msghdr		msg;
 	int			error;
-	size_t 			sent = 0;
+	size_t			sent = 0;
 
 	ASSERT(iop != NULL);
 
@@ -663,7 +663,7 @@ idm_iov_sorecv(ksocket_t so, iovec_t *iop, int iovlen, size_t total_len)
 	struct msghdr		msg;
 	int			error;
 	size_t			recv;
-	int 			flags;
+	int			flags;
 
 	ASSERT(iop != NULL);
 
@@ -981,7 +981,7 @@ idm_so_ini_conn_connect(idm_conn_t *ic)
 {
 	idm_so_conn_t	*so_conn;
 	struct sonode	*node = NULL;
-	int 		rc;
+	int		rc;
 	clock_t		lbolt, conn_login_max, conn_login_interval;
 	boolean_t	nonblock;
 
@@ -1324,7 +1324,7 @@ idm_so_svc_port_watcher(void *arg)
 	idm_so_svc_t		*so_svc;
 	int			rc;
 	const uint32_t		off = 0;
-	struct sockaddr_in6 	t_addr;
+	struct sockaddr_in6	t_addr;
 	socklen_t		t_addrlen;
 
 	bzero(&t_addr, sizeof (struct sockaddr_in6));
@@ -1386,7 +1386,7 @@ idm_so_svc_port_watcher(void *arg)
 		 * Kick the state machine.  At CS_S3_XPT_UP the state machine
 		 * will notify the client (target) about the new connection.
 		 */
-		idm_conn_event(ic, CE_CONNECT_ACCEPT, NULL);
+		idm_conn_event(ic, CE_CONNECT_ACCEPT, (uintptr_t)NULL);
 
 		mutex_enter(&svc->is_mutex);
 	}
@@ -2618,7 +2618,7 @@ idm_so_send_rtt_data(idm_conn_t *ic, idm_task_t *idt, idm_buf_t *idb,
 		mutex_enter(&ic->ic_state_mutex);
 		if (ic->ic_ffp)
 			idm_conn_event_locked(ic, CE_TRANSPORT_FAIL,
-			    NULL, CT_NONE);
+			    (uintptr_t)NULL, CT_NONE);
 		mutex_exit(&ic->ic_state_mutex);
 		mutex_exit(&idt->idt_mutex);
 		return;
